Sanitize, Escape, Validation, prefix on functions and class moved to different file.
