+++ rapaygo-for-woocommerce/readme.txt
@@ -2,9 +2,9 @@
 Contributors: claytantor
 Tags: lightning network, bitcoin, payments, payment request, woocommerce, automattic
 Requires at least: 5.7
-Tested up to: 5.9
+Tested up to: 6.0
 Requires PHP: 7.0
-Stable tag: 6.4.0
+Stable tag: 1.0.16
 Requires PHP: 7.0
 License: MIT
 License URI: https://opensource.org/licenses/MIT
